add tournament.signUpStart to view

+++ src/AppBundle/Controller/TournamentSignUpController.php
@@ -119,7 +119,7 @@ class TournamentSignUpController extends Controller
 
             $formDelete->handleRequest($request);
 
-            if ($formDelete->isValid()) {
+            if ($form->isSubmitted() && $formDelete->isValid()) {
 
                 //todo sometimes isUserRegister is null - error 500
                 $isUserRegister->delete();
